requirements.txt
adlfs==2025.8.0
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
azure-common==1.1.28
azure-core==1.36.0
azure-datalake-store==0.0.53
azure-identity==1.25.1
azure-mgmt-core==1.6.0
azure-mgmt-costmanagement==4.0.1
azure-monitor-query==2.0.0
azure-monitor-querymetrics==1.0.0
azure-storage-blob==12.27.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
branca==0.8.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
comm==0.2.3
contextily==1.6.2
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dependency-injector==4.48.2
dotenv==0.9.9
duckdb==1.4.0
executing==2.2.1
fastapi==0.135.1
fastjsonschema==2.21.2
fiona==1.10.1
folium==0.20.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
GeoAlchemy2==0.18.3
geographiclib==2.1
geojson==3.2.0
geopandas==1.1.1
geopy==2.4.1
greenlet==3.3.2
h11==0.16.0
h3==4.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
lark==1.3.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mercantile==1.2.1
mistune==3.1.4
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.3
objprint==0.3.0
osmium==4.1.1
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.3.260113
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
pmtiles==3.7.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pygeohash==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
pystac==1.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pywin32==311; sys_platform == "win32"
pywinpty==3.0.0; sys_platform == "win32"
PyYAML==6.0.3
pyzmq==27.1.0
query==0.1.4
rasterio==1.4.3
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.47
stack-data==0.6.3
starlette==0.52.1
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.41.0
viztracer==1.1.1
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
xyzservices==2025.4.0
yarl==1.22.0
